Main files for game
I've sent over the main class files for the project. There is obviously going to be way more code and subclasses, etc., I'm simply setting up what the baseline is for the project. Make sure when you edit, you comment in what changes/additions you've made to the code, that way we are all on the same page.